+++ flake8_assertive.py
@@ -25,7 +25,7 @@ import fnmatch
 import re
 
 __all__ = ['Checker']
-__version__ = '1.0.1'
+__version__ = '1.0.2'
 
 
 # Python 3.4 introduced `ast.NameConstant` for `None`, `True`, and `False`.
@@ -111,6 +111,16 @@ class Checker(object):
             yield self.error(node, 'A502', 'assertTrue', obj=True)
         elif any(arg for arg in node.args if is_constant(arg, False)):
             yield self.error(node, 'A502', 'assertFalse', obj=False)
+        elif any(arg for arg in node.args if is_function_call(arg, 'round')):
+            yield self.error(node, 'A501',
+                             'built-in rounding of assertAlmostEqual',
+                             op='round')
+
+    def check_assertalmostequal(self, node):
+        if any(arg for arg in node.args if is_function_call(arg, 'round')):
+            yield self.error(node, 'A501',
+                             'built-in rounding of assertAlmostEqual',
+                             op='round')
 
     def check_assertnotequal(self, node):
         if any(arg for arg in node.args if is_constant(arg, None)):
@@ -121,6 +131,7 @@ class Checker(object):
             yield self.error(node, 'A502', 'assertTrue', obj=False)
 
     check_assertequals = check_assertequal
+    check_assertalmostequals = check_assertalmostequal
     check_assertnotequals = check_assertnotequal
 
     def check_asserttrue(self, node):