+++ tests/test_checker.py
@@ -33,7 +33,7 @@ class TestOptions(unittest.TestCase):
 class TestChecks(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        # Always use Python 3's `assetRegex` method name.
+        # Always use Python 3's `assertRegex` method name.
         if not hasattr(cls, 'assertRegex'):
             cls.assertRegex = cls.assertRegexpMatches
 
@@ -66,6 +66,12 @@ class TestChecks(unittest.TestCase):
         self.check("self.assertEqual(False, a)", "A502", "assertFalse()")
         self.check("self.assertEqual(a, False)", "A502", "assertFalse()")
 
+    def test_assertequal_round(self):
+        self.check("self.assertEqual(1.01, round(a, 2))", "A501",
+                   "built-in rounding of assertAlmostEqual()")
+        self.check("self.assertEqual(round(a, 2), 1.01)", "A501",
+                   "built-in rounding of assertAlmostEqual()")
+
     def test_assertnotequal_none(self):
         self.check("self.assertNotEqual(None, 1)", "A502", "assertIsNotNone()")
         self.check("self.assertNotEqual(1, None)", "A502", "assertIsNotNone()")
@@ -83,6 +89,16 @@ class TestChecks(unittest.TestCase):
 
     def test_assertnotequals(self):
         self.check("self.assertNotEquals(True, a)", "A502", "assertFalse()")
+
+    def test_assertalmostequal_round(self):
+            self.check("self.assertEqual(1.01, round(a, 2))", "A501",
+                       "built-in rounding of assertAlmostEqual()")
+            self.check("self.assertEqual(round(a, 2), 1.01)", "A501",
+                       "built-in rounding of assertAlmostEqual()")
+
+    def test_assertalmostequals(self):
+                self.check("self.assertEqual(1.01, round(a, 2))", "A501",
+                           "built-in rounding of assertAlmostEqual()")
 
     def test_asserttrue_is(self):
         self.check("self.assertTrue(True is True)", "A501", "assertIs()")